rag-document-assistant/requirements.txt
endee
sentence-transformers
langchain-text-splitters
streamlit
PyPDF2
python-docx
striprtf